+++ lms-backend/src/test/scala/tests/it/lms/LmsLinAlgItTests.scala
@@ -28,58 +28,58 @@ class LmsLinAlgItTests extends ItTests {
   
   import progSeq._
   
-  private def sparseVectorData(arr: Array[Int]) = ((0.until(arr.length)).toArray, (arr, arr.length))
+  private def sparseVectorData(arr: Array[Double]) = ((0.until(arr.length)).toArray, (arr, arr.length))
 
   test("ddmvm") {
-    val inM = Array(Array(1, 1), Array(0, 1))
-    val inV = Array(2, 3)
+    val inM = Array(Array(1.0, 1.0), Array(0.0, 1.0))
+    val inV = Array(2.0, 3.0)
     val in = Pair(inM, inV)
-    val out = Array(5, 3)
+    val out = Array(5.0, 3.0)
     progSeq.ddmvm(in) should be(out)
     lmsTestRun(progSeq, progStaged)(progSeq.ddmvm, progStaged.ddmvm)("ddmvm", in)
   }
 
   test("dsmvm") {
-    val inM = Array(Array(1, 1), Array(0, 1))
-    val inV = sparseVectorData(Array(2, 3))
+    val inM = Array(Array(1.0, 1.0), Array(0.0, 1.0))
+    val inV = sparseVectorData(Array(2.0, 3.0))
     val in = Pair(inM, inV)
-    val out = Array(5, 3)
+    val out = Array(5.0, 3.0)
     progSeq.dsmvm(in) should be(out)
     lmsTestRun(progSeq, progStaged)(progSeq.dsmvm, progStaged.dsmvm)("dsmvm", in)
   }
 
   test("sdmvm") {
-    val inM = Array(Array(1, 1), Array(0, 1)).map(sparseVectorData)
-    val inV = Array(2, 3)
+    val inM = Array(Array(1.0, 1.0), Array(0.0, 1.0)).map(sparseVectorData)
+    val inV = Array(2.0, 3.0)
     val in = Pair(inM, inV)
-    val out = Array(5, 3)
+    val out = Array(5.0, 3.0)
     progSeq.sdmvm(in) should be(out)
     lmsTestRun(progSeq, progStaged)(progSeq.sdmvm, progStaged.sdmvm)("sdmvm", in)
   }
 
   test("ssmvm") {
-    val inM = Array(Array(1, 1), Array(0, 1)).map(sparseVectorData)
-    val inV = sparseVectorData(Array(2, 3))
+    val inM = Array(Array(1.0, 1.0), Array(0.0, 1.0)).map(sparseVectorData)
+    val inV = sparseVectorData(Array(2.0, 3.0))
     val in = Pair(inM, inV)
-    val out = Array(5, 3)
+    val out = Array(5.0, 3.0)
     progSeq.ssmvm(in) should be(out)
     lmsTestRun(progSeq, progStaged)(progSeq.ssmvm, progStaged.ssmvm)("ssmvm", in)
   }
 
   test("fdmvm") {
-    val inM = (Array(1, 1, 0, 1), 2)
-    val inV = Array(2, 3)
+    val inM = (Array(1.0, 1.0, 0.0, 1.0), 2)
+    val inV = Array(2.0, 3.0)
     val in = Pair(inM, inV)
-    val out = Array(5, 3)
+    val out = Array(5.0, 3.0)
     progSeq.fdmvm(in) should be(out)
     lmsTestRun(progSeq, progStaged)(progSeq.fdmvm, progStaged.fdmvm)("fdmvm", in)
   }
 
   test("fsmvm") {
-    val inM = (Array(1, 1, 0, 1), 2)
-    val inV = sparseVectorData(Array(2, 3))
+    val inM = (Array(1.0, 1.0, 0.0, 1.0), 2)
+    val inV = sparseVectorData(Array(2.0, 3.0))
     val in = Pair(inM, inV)
-    val out = Array(5, 3)
+    val out = Array(5.0, 3.0)
     progSeq.fsmvm(in) should be(out)
     lmsTestRun(progSeq, progStaged)(progSeq.fsmvm, progStaged.fsmvm)("fsmvm", in)
   }
